+++ CHANGES.md
@@ -1,8 +1,9 @@
 # Changes
 
 - 0.0.4: When listening for new controller input: 
--- Fixed a bug with the cmd file for starting the jar
--- Character movement: Map point on square to point on unit circle for joysticks
+    - Fixed a bug with the cmd file for starting the jar
+    - Character movement: Map point on square to point on unit circle for joysticks
+    - Mouse movement: Add a setting for the stick threshold
 
 - 0.0.3: Automated deployment via GitHub Actions
 - 0.0.2: Update deployment and versioning